Samurai patch
NOTE: the extern.h entry for roleSkill() was included in the previous commit

Female Samurai are started with a glaive and a knife instead of a katana and a shortsword.

The Samurai bonus with glaives was put into the same framework as the skill bonuses from binding spirits. As a consequence, the bonus will be reflected on the #enhance menu, if you look at that menu while wielding a glaive.

Samurai can reach expert in darts. Wizards can only reach skilled