Pirate Patch: Reaver modified
Reaver was a pretty major source of inventory clutter. It now asks players if they wish to take the item it has just stolen. If they opt not to take the item, they can pick a direction to throw it in or choose no direction to drop it at their feet.