Noble Patch
Adds the Noble role to the game. Humans, incantifiers, and vampires
can be Nobles. Nobles are a melee-focused role that specializes in
one-handed weapons. They can engage in some spellcasting if their
ability scores are good enough.

Nobles have an automatic first gift, the Rod of Lordly Might.

Their quest artifact is an artifact opera cloak (vampires) or ornamental
cope (non-vampires). Note that this means that the base type of the
Noble quest artifact changes from game to game.