extern.h
Update list of functions

artifact functions should reflect that artifact IDs are ints now.

dotravel is now used by Thought Travel on Astral

Evaporation of the Armor plates on Noble's dresses can change the slot the dress is equipped in, need to call shirt functions to do that

Add function for telling if the current level is a cave

Add a 'reset hunger' function, so that the initialization code in init_hunger isn't run after near death experiences.

Add function for selling clockwork components (ask_cp)

Add function for determining allied drow factions

Add function to handle LOUD sources of monster waking (for deafening grimlocks)

Add mutator function for the bypassDR variable

Add function to allow drow to join their nemesis

Add function to tell if a quest art is the original or a traded version

Drow need to be able to use set_lit

Too-hit functions work a bit different now, taking and generating values for regular, touch, and physical

illurien does not permanently die.

The Claws of the Rev revive eggs.