potion.c 3.8
Clockworks that have furnaces need potions of water.  The food is burned in the furnace and boils the water.  Up to MAX_BOILER water can be saved

Potions of Haste give permanent speed

Blessed Potions of Gain Energy completely refil your energy

Poitons of Oil are burned in the oil stove upgrade

Potions of deep dragon or newt blood can serve as gain energy in alchemy