Library Patch
Adds a new special room to dnethack, the waterlogged library

Embeds spellbooks in the walls of the room (representing shelving), most are blank, but some are random spellbooks.

Has 1-4 living lecturns in it.  Living lecturns are a type of wood golem that can shoot Magic Missiles, and leave a random spellbook behind when they die.

Also has pools and occasionally fountains.

Found on dungeon level 9 and below