Properly declare some pronoun functions
This fixes a crash on some things using himherit and a few related
functions, infamously, the Echoes of the Last Word (Tenebrous) and
Open (Otiax) spirit powers.

C assumes an undeclared function to return 'int' by default for
historical reasons. If the function actually returns a char*, things
will proceed to blow up ... or not: on some architectures, int and char*
are the same size and treating an int as a char* Just Works(tm).
The code puts its fingers in its ears and goes "la la la I can't hear
you, this peg fits into that slot so I'll just treat it as a pointer".
On other architectures, sizeof(int) < sizeof(char*). OOPS.
...and that's why Chris never managed to reproduce the bug,
as far as I can tell. Blame the C language. :)