4 fiends patch
The 4 fiends are now waitforyou, don't need to be asleep

The Flare spell needs a big explosion.  Also, enemies that cast it should be immune to explosive force (blaster bolts, flare spell, and I think that's it?)

Implement FF magic spells.  Have I mentioned that ranged weapons are probably critical to this game? 'Cause ranged weapons are probably critical to this game.
	Revise Fiend spell lists to match FFI mechanics
	Add concept of an AoE spell, that can still be cast at the wrong square.

Killing Garland and the fiends results in the crystals being dropped by whichever is killed second.

Make Chaos 50% resistant to wards now that he doesn't get spellcasting