Fix warnings in apply.c
This brings in pline1() and friends from NetHack 3.6.0 to suppress a
number of format-security warnings.

This also fixes use_sensor() to actually take a turn when it's used,
nixes a couple of dud expressions in the first part of a couple of
for-loops, and supresses a warning where an assignment happens in an if
predicate.