Fix warning in display.c, segfault in #dump_map wizmode command
There was a '\0' in the format string for the #dump_map wizmode command
which did nothing except raise a compiler warning.

In the course of testing, I also discovered that the #dump_map logic
used a buffer that was too small to do its task without segfaulting, so
I made it large enough (5 chars * 80 columns == 400 bytes).