Fix warnings in do_wear.c and armor-charisma interaction
Fixes a '/*' in a block comment and an implicitly used function.

Also fixes adj_abon() adjusting charisma for *all body armors* instead
of limiting the effects to the noble's dress, consort's suit and
gentleman's suit (under-shirts were unaffected).

Gets rid of a duplicate check for gentleman's suit which was there for
some reason.