Fix warnings in dungeon.c
* Include ctype.h so toupper() is declared.
* boolean -> BOOLEAN_P and xchar -> XCHAR_P in declarations.
* Format security.
* Parenthesis disambiguation.

There was one case where the *address* of In_cha() was being used as a
condition instead of actually calling In_cha().  It doesn't change
current behavior, but it would be a weird source of bugs should
In_outdoors() ever be extended; it's only used to check if the Simurgh
can be drawn.