Fix warnings in shk.c, handling of basic/premium shopkeeper ID service
Fixes a couple of non-void functions that should have been void, and a
handful of format security warnings.

Also fixes the check for availability of [b]asic and [p]remium ID from
shopkeepers; it was based on multiple bitwise operator ambiguities
defaulting to an expression that was almost certainly incorrect.