Fix warnings in spell.c, add missing letters to spirit power prompt
Fixes a few implicit function uses, an unclosed block comment,
char -> CHAR_P in a function declaration, implicit returns, assigns in
if tests, format security warnings, and %d -> %ld for long int format
specifiers.

Also adds letters to the sprit power prompt which were supposed to be
visible but weren't due to a mistake in the format string.