Fix warning in wield.c
Just a && and || disambiguation, but it took a long time for me to
figure out that dNetHack made aklyses, Sol Valtiva (a two-handed sword
artifact) and The Pen of the Void with the Marionette seal like polearms
with reach, but also allows them to be used in melee.  I took the
liberty of reindenting the bash-weapon flag expression to clarify what
the hell was going on.