Fix warnings in worn.c
Function defined without a return type; I made it void because it wasn't
supposed to return anything anyway.

Also, clang complains about a constant used in with a logical operator.

"warning: use of logical '&&' with constant operand
[-Wconstant-logical-operand]"

This is actually a false positive: clang thinks the constants are
superfluous, which they would be if they were actually written manually,
except they're not, they're macro expansions.  gcc handles this
correctly by treating the macro invocation more like a function call and
not raising the warning, which matches what the programmer actually
meant in this case.