Rnl works better with large input.
Some of these are still problematic, I think, since there are comparisons to values within low double digits of the end of the range?  But at least it shouldn't suddenly go to .5% or whatever at -2 luck.

This changes the effect of SEEK/eyeglasses/Otiax on searching.  I'd also like to make searching WIS-dependent