Pet management: #wait and #come
Add extended commands to prevent one or all pets from following between levels, or order one or more to resume following between levels.

Commands take no time for ease of use.  Commands use a direction prompt for ease of use, but this may change if it proves anoying in play.

#wait command times out after 100 turns, to prevent pets from remaining behind due to a forgotten command.