Fix major pet AI bug
Pets use mtrack (just like monsters) to cut down on
oscillating/random movement. However, for pets, this is
detrimental, since it means that if the player lingers, or the
pet is faster than the player, the pet might end up going in the
wrong direction and get "stuck" going the wrong direction. This
is why pets basically never follow the hero properly in
NH3+variants.