Gnome quest staircase adjustments
Turns out the branch staircase code is easily flexable enough to support regular branch staircases connecting the quest home to minetown.  So just do that instead of special-casing everything.

This results in a post-level-gen modification to the minetown map, and retains a special-case handling the quest home staircase.

This also retains the special "fall through hole" case, even though it's no longer needed.