Fix wishing related to "holy", "unholy"
Bug fixed: wishing for anything "holy" was broken and would fail
Bug fixed: wishing code for "holy" and "unholy" weapon props would not be reached

Patched behaviour:
In addition to weapon properties, wishing for a "holy" item will also act as wishing for a "blessed" item, unless specified otherwise prior in the string. Same for "unholy", and "cursed". Thus, wishing for an "uncursed holy dagger" will yield a holy-property dagger with an uncursed BUC.
Note that wizard-mode wishes will force object properties, so a wizard-mode test of wishing for a "holy spellbook of remove curse" will have a different result for a real wish with the same wording.