Refactor dmgval()!
dmgval_core() is able to describe the damage dice of most of dnethack's weapons

Although the switch(otyp) in dmgval_core() is equally unreadable to the switch(otyp) that used to be in dmgval(), it can be fed into other functions (like item descriptions!) to be very readable. It's also fairly easy to manipulate once you've stared at the #define mess for a while.