Items made of gemstone are specific kinds of gems
Stores gem type in ovar1, so items/artifacts that use ovar1 are ineligible.

The type of gem is known based on the knowledge of the real gem - identifying one will reveal the other:
ie, rubbing a "pink gem dagger" on a blessed touchstone would reveal it to be a "morganite dagger" and reveal the pink gem in your inventory to be morganite at the same time.

Can be specifically wished for (following existing re-material wishing rules).

Not accepted by jewelery stores as of yet, since nothing related to materials and pricing has been done yet.

Also, misc small improvements to related functions (material_name and the wish parser).