Use less realistic material densities
Copied from xnh, with minor changes.
Bronze is the same as iron, dragonhide and paper are slightly heavier, mithril is lighter.
Obsidian is the same weight as glass, shadowsteel is the same weight as wood.

Adjust many items' weights.
Elven equipment is less incredibly light.
Many other pieces of armor are lighter.