Material affects armor defense
Note that unlike (almost?) every other variant that has this, the effect is multiplicative, not additive. This means that low-stat armor (which is everything outside of the body-armor slot) tends to see little to no change in effectiveness. For example, mithril (rating 6) gauntlets are equally effective as iron (rating 5) gauntlets.
This method was chosen for a couple main reasons: Existing mithril equipment (high elven gauntlets + helm, for example) was no better than their iron counterparts AC/DR-wise -- making rematerialed items behave otherwise would either be inconsistent or require (in my opinion, unnecessary) buffs to these items and others in similar situations. It allocates defense bonuses along the existing AC/DR split of each item. Many accessory armor pieces deliberately have little to no contribution to AC, but mostly act as DR for their slot.
Still, elven plate, droven plate, and droven chainmail were buffed to reflect them having a superior base material.
To prevent shenanigans, the cloak of protection's great AC/DR for its material was changed to be a flat +2/+2 in arm_<x>_bonus.

Crystal equipment was buffed:
Has equal base protection to standard iron, but only gains a 1/3 bonus to enchantment (ie +7 acts as +9) (old behaviour was that +7 acted as +14).
Weighs significantly less, approx as much as the standard iron version made out of glass (ie 3/4 the weight of iron). Who in their right mind wore the 300wt not-flying not-speed crystal boots, or the 150wt can't-drink-potions crystal helm for a few points of defense? If this turns out to be too much of a buff, we can adjust it again but I doubt there will be a ressurgence. CPM still weighs more than dragon scale mail for +2/+3 AC/DR without the properties or half-dex-AC.