Change how wishing for object materials works
Limited to what could randomly generate (eg, no liquid daggers)
Probability of material being granted is increased if that material was not rare to begin with -- for example, iron weapons have a 5% chance to generate as silver, but since there are another 10 percentage points of rarer materials than silver, wishing for a silver dagger has a 15% cumulative chance.
Furthermore, there is a 1/7 chance that the above probability check is neglected, and if it's possible for an item to be of that material, it will be -- for example, leather gloves have a 1% chance to generate as dragonhide, but wishing for dragonhide gloves will give dragonhide (barely more than) 1/7th of the time.
When wishing for an artifact, materials are only allowed if that artifact is nameable or Excalibur.
Wizard-mode wishes ignore the above logic.

Also, bug fixes in wishing parser.