Fix vivisection "trap" behavior when restoring level, add error-recovery code.
The fix changes the order in which events occur when reloading a level.  This probably won't cause new crashes :/

The error recovery code will trap anything that ends up on a vivi trap after a level is reloaded, which seems bug prone.  The code should therefore be removed after a version or two.