Monster equip improvements
The player can no longer informally-ID amulets of life saving and reflection by trying to equip a pet with them. The amulet type must be known. More amulets are allowed via #equip, though monsters will still only actively put on life saving and reflection.

A worn amulet of null magic will stop a monster from casting spells. Currently, the only way to forcibly put an amulet on a monster is if said monster is a pet, so this should be fine.

Update the function for monsters' favourite equipment. Monsters will prioritize MR and reflection and speed, and generally like other properties.