Backend changes for Arcadian Forest outposts
Added to the shoplist proper.
Changes the symb in the shop structure to shoptyp; places where a symbol is actually needed can be done by accessing the first item index of the shop, which is klunky, but far more often we want to know exactly what shop we are working with.