fixed the bugs As a half-dragon, wearing the appropriate dragon scale armor (or shield) will boost your breath attack. Each armor piece adds 0.5x (max 2x damage for shield and scales/mail). The chromatic dragon scales will give a bonus to fire, cold, shock, acid, and poison breath (anything but sleep basically), and the platinum dragon plate will give a bonus to fire, cold, shock, sleep (not acid or poison). This lines up with the resistances they grant. I didn't bother implementing this for black, silver, or gray since I don't forsee those being valid half-dragon breaths anytime soon.
Without any boost, half-dragon breath was equal to magic missiles. With a full 2x multiplier, it's equal to magic missiles with double spell damage.

Note - the dragon's heart-stone will give a flat 2x modifier that _stacks_ with this, for a theoretical max multiplier of 4x. This comes out to 4*16d6, avg 224 damage / max 384 damage