Allow basic wishing for "petrified" statues, which attaches a monster to the statue
This creates a monster of the appropriate corpsenm to attach to the statue, which is needed for statue traps to generate a unique monster. This is NOT enough to specify a statue of a derived undead or tame creature.
Use this to attach the Shaktari monster to the statue in Demogorgon's lair.

Also works for corpses.