black web allows 2wep with blank offhand
Even if you can't normally two-weapon, you can two-weapon any one-handed weapon with a shadowblade or two-weapon. This obeys the standard two-weaponing rules, and does NOT let you two-weapon a blank main hand with an offhand weapon. The latter should rarely matter but if there's some item with a penalty for wielding it, it won't bypass that I guess.