Beartrap in shop setting message
Gives a message now when you set a beartrap up in a shop if it's not yours, instead of just silently adding it to your bill. I also added a message if its yours just for completeness.

Note - it's a stupid idea to kept placing / setting traps, since you'll make money for every time you untrap it (and sell it back to the shopkeeper), but you'll pay the full cost when you place the shopkeeper's trap done, so it's a losing battle unless you've got 25 charisma, where you just break even.