Glyphs and HP and Energy changes
Change how HP is calculated to allow greater control

Change the length of the healing cycle to 90 turns from 30, and add a constant to make it easier to change in the future

Implement glyphs

Change how Energy is calculated to slightly reduce dependence on Wis

Dracae don't spawn enemies if they're peaceful.

Do extra damage vs. enemies with one of the "vulnerable" flags if their vulnerable damage type is used.