Adjust the mad_turn algorithm to use the single-hashed turn number and the double-hashed turn number
Hopefuly the distribution of this will be closer to the two-random-number distribution rather than the random-number-squared distribution.

The empirical distribution should probably be measured.

Also could use the addition of a per-session (or even per-turn) nonce to reduce or remove patterns in the order in which effects trigger (adding the hash of the madness ID has hopefull already mostly removed this problem).