add bestial claws
very similar to the beast claw from bloodborne

d10/d8 base damage, uses bare handed combat skill (but is a wielded weapon, so gets no bonuses from gloves or anything like that). without the beast's embrace glyph it's just that with no other special modifiers.

with the beast's embrace glyph (which is generated with a 1/4 chance on every werewolf in the depths, so you've got around a 40% chance to have at least one, it gets pretty big bonuses. the beast claw itself does sneak attacks. it doesn't get any other main hand bonuses. however, if you're wielding nothing in your offhand you can two-weapon (even if you can't two-weapon) to add a pretty great claw attack (which takes the place of your offhand weapon attack).

the claw attack is just AT_CLAW AD_PHYS. the number of dice is <lvl/8+1>, so 1-4, hitting 4 at xp 24. the die size is based off your current insight level. unlike the other two insight-related weapons in the game, you want minimal insight. the die size is (spe + ((60-insight)/20)^2). that comes out to spe + 5 at insight 15, spe+1 at insight 40, and spe+0 for insight 60. it can't go lower than max(spe, 1). as i write this commit message i realize i need to change it to use float division not integer division, so that'll be addressed in the next commit.

at level 30 with 15 insight and +7, the claw will deal 4d11 damage, in addition to the d10/d8 + 7 from the main hand weapon attack. that's pretty good damage i think.

it's worth noting that the claw + sneak attack both requires the glyph to be ACTIVE. insight level is set to 15 for that glyph, so it won't ever trigger if insight<15, just like any of the other glyphs I think. something like that.

bestial claws can be found on any werewolf in the depths. they also now start with a helmet and leather armor/gloves/boots, in a poor approximation of the bloodborne yahar'gul black set that the dudes wielding them wear. to make sure at least one will spawn, one of the 3 random monsters that was placed in alhoon 1's level is a werewolf (it was just a completely random spawn in the des file and didn't look like it mattered much), and the iron golem in the alhoon fort is changed to a werewolf as well (since that predated the alhoon being there I figured it was fine?). there is technically no guaranteed glyph but i figured that was fine - there's no guaranteed rakuyo either.