makemon() should always place monsters safely
No matter what combination of whether or not ptr is given or whether or not coords are given, makemon() should check that a monster will be okay for its location.
This removes the need for the MM_CHECK_GOODPOS flag.

This also unearthed a bug -- directly setting fakemon.data is not sufficient, because many aspects of monsters are now governed by their properties (most noticably, phasing) which was causing monsters to not get placed in spots where they should have been able to go.
This commit fixes said bug. Xorns are once again able to spawn in solid rock!