kusanagi can be picked up safely
Previous behavior: if you named the sword then dropped it, it got to sit on the floor forever, since it wouldn't let you pick it up.
Now, if you're capable of naming the sword, it won't evade your grasp. If you're not capable of naming it, it will still blast and evade.

It specifically won't let you wield it unless you fulfill the criteria. If you get drained while wielding it, you'll be fine, you just can't re-wield it until you cure your levels. This won't work well if there's any other way to wield a weapon that bypasses ready_weapon in wield.c