Fix unfree ball crash bug
The player could throw their ball into the inside of an engulfing monster, which adds the ball to the monster's inventory.  When expelled, the ball was therefor not able to be placed.

Check if the ball is being carried by a monster, and if it is extract it before calling placebc()