Add give_intrinsic function to unify things
Unifies all of the copy-paste for giving the player a temporary intrinsic into one function. `give_intrinsic` takes 2 arguments - the property (FIRE_RES/etc.) and a long for the duration. If the duration is -1, it's permanent, and if the duration is -2 then that intrinsic is removed.

Calls incr_itimeout. It's possible that the incr_itimeout function should be consolidated and parts of them merged, but I don't know why they were separate in the first place, so they're left alone for now.