Add monster tip interactions
Can tip a rothe to wake it up, but knocking it over. Can tip a shk to offer them gold. Tip a leprechaun to bribe them to go somewhere else, pacifying them. Tip a migo philosopher to offer them a penny for their thoughts. Tipping a mercenary will prompt to throw gold if you want to bribe.