Allow wishing for figurines of uniqs/nowish monsters
The attempt to create the monster will just fail, if it would have previously been nowish. Is this rude to players who wished for a figurine of <x>? Yeah, but they wouldn't be getting the figurine anyway. The intention is that if you really want a swag figurine of demogorgon to ascend with, then by jove you'll get one!