Fix ls bug
Steps to reproduce bug:
- Turn on monpolycontrol
- ^G a chameleon into yellow light form
- let it die

On death, the chameleon's data is set to `&mons[PM_CHAMELEON]`.
m_detach() was checking mon->data "should this monster emit light".
It was not considering mptr, which would have said "this monster used to emit light".

TODO:
Actually store ls pointers (and timers?) on monsters and objects instead of in the processing chain, so that we can more easily check "does X have an attached lightsource"