Adjust bonus damage
It's now a flat +mlev bonus, but reduces the remaining time by (mlev/10) - rn2(5) (floored, never increases duration). That means that you chould hypothetically get an average of 4 or so hits on Demogorgon, each dealing +108 flat in addition to whatever base damage it deals. Stupid? Yeah, quite possibly. Reasonably balanced? Dunno. I'm expecting it to be quite lethal at killing demon lords if you bother to ward & prep beforehand, but not otherwise. Might revisit the initial damage and increase it to 40% or something like that.