Fix drowning issues
Before instant death, drowning first uses up your breath meter. This puts you underwater as though you were amphibious. Moving next to a crawl-able escape instantly and immediately brings you to shore (as before). This may need to be adjusted in the future for balance / verisimilatude.

Teleports do not consider 3D-water to be safe if you can only survive in it by holding your breath.

Fixes infinite drowning-loop on the plane of water.