Fix xpathto's memory allocation
It allocated an Cx(R+2) length block of memory:
```
Ex: C=8, R=2. This is one continuous block; the end of one line goes to the start of the next
000000000
000000000
000000000
000000000
```
Then would loop over an inner portion to mirror the game map, of size (C-1)xR:
```
Ex: "l" is where is being looped over
000000000
0llllllll
0llllllll
000000000
```
And would check if one square is "adjacent" by adding/subtracting the right amount (+1, -1, +C, -C, +C+1, +C-1, -C+1, -C-1).
In *most* places, this would cause it to correctly access one of the designated non-looped portions (marked by a 0 in the examples).
In one place, this would cause it to access unallocated memory:
```
Ex: X is the looped-over spot, and Z is the unallocated memory being accessed
000000000
0llllllll
0lllllllX
000000000Z
```
Easiest fix: allocate that one more byte of memory.

Also, improve the commenting of that function.