Revert changes made to remove_rooms(); use different Spire-Dispensary fix
Turns out I made a big mistake changing remove_rooms().
The roomno of parts of levl[][] wasn't being cleared when rooms were.

This rolls back remove_rooms() to prior to that change, which reintroduces the issue of the dispensary stairs appearing behind the Spire.
Fixes that issue by forcing the dispensary stairs to appear in one of the prior levels (it can never appear on the spire level now).