Improve "player" monster equipment.
Add init_mplayer_gear() to ensure that they start with full sets of representative gear.

This moves some of the code from father down in the mplayer code up to a dedicated function.

Needs an "upgrade" function to better handle improving on their gear for the astral plane mplayers while still doing something to maintain theme.