Fix lightsource radius code.
The highest radius is 10*3: 30

Adjust MAX_RADIUS, circle_data, and circle_start[]

Note: Info in circle data calcualted based on circles with sums-of-squares of r^2+r, where r is the radius of the lightsource.

Add note to mondata.h explaining the connection between emits_light and circle data et al.