ID and mitigate potential corruptions of monster movement loop
If a monster's actions cause the next monster in the fmon chain to be moved to the migrating_mons chain, the `nmtmp` pointer will cause the loop to go through the migrating_mons chain -- this is BAD.
Check that each new mtmp in the loop is on the level (by checking mx and my).

If the monster is migrating, we cannot use its `->nmon`. We also cannot trust the previous mtmp to still be on the level, even if we had a pointer back to it. For a perfect solution, we would need a list of every monster handled by the loop so far.
Instead, we will restart the loop at fmon again.
This will double-process any monsters that were already handled AND have >=12 movepoints remaining, letting them get their second move earlier than expected (in relation to other creatures' movements) this global turn. I think that this is minor enough to be unnoticable, especially with how rare the trigger event is in the first place.