Use a leftover bitfield bit to track which monsters were *created* in the afterlife.
Such monsters don't leave corpses (unless they are riders)

This doesn't account for the fact that most of the alignment quests are technically *also* set in the afterlife, but many of the upper planes have lots of living mortals on them and the alignment quests focus on such areas.

I tested to make sure my bit-count was accurate and that this didn't change the size of the monst struct.