Close lockfile properly
Presently dnh leaks handles to the livelog file from every event. This
occurs because the lockfile (which is the same as the file itself in the
case of the livelog file for some reason) never gets closed.

This was noticed for vanilla as well and fixed in vanilla commit d1391149
This commit is a copy of that fix