Safer clone_mon()
Looked like it was possible for a tame minion cloned on your turn to fail the remain-tame check and thus not copy the EMIN structure from mon to m2 while preserving m2->isminion.

Honestly, minion and dog code needs a lot of cleanup because the original made a *lot* of assumptions. This is a start.