Rework Oracle, adding enlightenment and dungeon inclusion hints
Also migrates the dialogue to a menu, instead of the series of y/n/q it was before.

Enlightenment retails for 200 + 20 * level, and works identically to a standard wand zap.

Glimpses of the future gives a random hint about the various set-on-game-start inclusions - Gae vs. Tulani spawns, Dracae vs. Polypoid vs. Tulani spawns, the Platinum Dragon's airship, your
chaos variant, and Oona's elemental affinity. Did my best for vaguely flavorful messages for each option, trying to have fun with it, but I'm not super set on any of the messages.