More flags for summons
Unused as of yet, but allow summons to be flagged to never try to change floors.
Use a proper "summon will not time out" flag.
Allow summons that have no summoner to travel between levels (if not explicitly denied via flag when being marked as summoned).

Also, fix a couple of necronomicon bugs.