Desummon_obj needs to handle obj->ocarry's weapon pointers. Because obj_extract_self() doesn't, yuck.
It would be quite inefficient to add to obfree() the same pointer-catching that is done for the players' equip slots -- yes, let's loop through all monsters to see if one of them was wielding this weapon. No we can't use obj->ocarry to avoid looping through all monsters because obj's location was set to OBJ_FREE when it was extracted.