Prevent cursed offhands slipping when weldproof
When starting twoweaponing with the offhand cursed, or when the
offhand weapon becomes cursed whilst already twoweaponing, it no
longer drops to the ground.

The only case not considered here is losing the weldproof effect whilst
already twoweaponing, by e.g. curing lycanthropy

One can imagine that you already have a really tight grip on the weapon
knowing that you'd otherwise lose it - this is the little-known strategy
'using your imagination to avoid fixing edge cases'