Special case the artifact iron ball at pickup
The artifact iron ball of levitation actually improves your carrycap by
its weight whilst carried.

If you have limited carry capacity for any reason, say you're a gnome
or you've read scrolls of punishment, you may not be able to pick it up
to benefit from that effect.

This special cases the check in the number of objects you can pick up,
as well as the guard for warning about encumberance