Hell Vaults/Hellish Seals
Seals are terrain. Digging down into one spawns a miniboss and burries a treasure chest with magic gear, including oprop gear, under the spot where the seal was.

The monster type is pre-chosen and the broad class is indicated by a statue.

However, because the monster doesn't exist untill it attacks it will hopefully be resistant to cheesing.