Add the veil
Most characters start the game with the veil intact.

Characters with an intact veil don't gain insight, and gain a +10 bonus to sanity regeneration.

Entering the depths removes the veil, as does successfully reading the Necronomicon and Book of Lost names.

Binders, Madmen, and Anachrononauts begin the game without a veil.

Makes entering the depths slightly more fancy.