mofflin() etc. bugfixing
mofflin() caused a crash by returning negative coordinates as good.
On examination, it appears that a backwards check will cause bad behavior
on nearly full levels, by 1) returning the un-initialized contents of [0]
if no good spots were found, and 2) by failing to move the monster if only
a few good spots were found.

Also, check what the find-coordinate family of functions is attempting to
return and do an impossible() if it's actually invalid.