Tobiume now tases enemies
Tobiume still deals +1 fire damage, but the on-death explosions and invoke now deal a combination of fire and shock explosions. It also gives both fire and shock resistance when wielded, up from just fire.

If the unskilled invoke is used (single beam), the damage type is chosen randomly. If the skilled invoke is used (cluster of explosions), the damage type is chosen randomly and independently for each explosion.>
For the on-death explosions:

* If the monster resists fire but not shock, it will use a shocking explosion
* If the monster resists shock but not fire, it will use a fiery explosion
* If the monster resists neither fire nor shock, it will use either one with a 50% chance each
* If the monster resists both, it will use a fiery explosion

Also fixes a minor missing period if the player does not have enough energy to use the shikai.