Specialize vault items some
Items that are going to be kept (artifacts and magic armor) have a much higher chance of special properties.

After the one-in-ten check for a random major property, a new 1-in-4 check is made for a themed property.
-Holy vault
--items are blessed
--may be made of silver or gold
--may give fire, cold, elect, or reflection.
--may have fire, cold, elect, holy, any alignment, wrathful, or vorpal damage

-Devil vault
--items are cursed
--may be made of gold or green steel
--may give fire, cold, elect, or magic resistance
--may have fire, cold, elect, unholy, axiomatic, life draining, or flaying damage

-Demon vault
--items are cursed
--may be made of gold or iron
--may give acid fire, cold, or magic resistance
--may have acid fire, cold, magic, unholy, anarchic, life draining, or flaying damage

Add wrathful, life draining, and living weapons to the list of generic special properties.