`makemon()` can now directly specify template and faction to apply
Actually, `makemon()` behaves exactly as it did before, but I've added another version, `makemon_full()`, where those things are specifiable.

Things that affect monsters being of a template (other than directly being specified as such by `makemon_full()`) belong in `makemon_set_template()`.
Things that... ...faction... ...belong in `makemon_get_faction()`.

This was as much of a kerfuffle as it was because monster placement will take templates into consideration. If we were to add an "atlantean" template that gives monsters waterbreathing, and called `makemon_full()` with it (or the level defaulted to putting that template on monsters), the monster will accept being spawned in water. This was not possible (as cleanly) before.

I was able to kill `mkmon_template`, but `curhouse` is needed for peaceminded(), and I didn't want to try to dig into `mkzombie()` for `zombiepm` and `skeletpm`.