Futureproof: set template shoouldn't check mtemplate_accepts_mtyp
Since the preceeding code picks random templates according to the rules, this would only catch mkmon_full being told to apply an unusual template. The desired behavior there would be for makemon to apply the template without comment.

An example might be if Orcus ends up with zombie angels in his employ.