New artifact: "Mask of Many Faces"
Neutral mask
Invoke while wielded or worn against a target adjacent creature.
If the creature is below 1/3rd hp (1/6th for uniques), kills them leaving no corpse. The Mask then assumes the shape of the slain monster.
When worn (or immediately, if it was invoked while worn), the player is polymorphed into the form of the Mask, including for no-poly monsters.
The duration of the polymorph is dependent on the relative level of the player to the monster, reduced significantly for no-poly monsters. Unchanging cannot be used to make the effect permanent.
While poly'd, the mask can be re-invoked to end the effect early.